requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-retry==2.9.1
aiosignal==1.4.0
altair==6.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.96.0
anyio==4.13.0
attrs==26.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
blis==1.3.3
cachetools==7.0.6
catalogue==2.0.10
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cloudpathlib==0.23.0
colorama==0.4.6
confection==1.3.3
cryptography==46.0.7
cymem==2.0.13
distro==1.9.0
docstring_parser==0.18.0
et_xmlfile==2.0.0
fastapi==0.136.0
filelock==3.29.0
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.47
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.12
Jinja2==3.1.6
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.15
langchain-core==1.3.0
langchain-openai==1.1.16
langchain-pinecone==0.2.13
langgraph==1.1.9
langgraph-checkpoint==4.0.2
langgraph-prebuilt==1.0.10
langgraph-sdk==0.3.13
langsmith==0.7.33
lxml==6.1.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.1
murmurhash==1.0.15
narwhals==2.20.0
numpy==2.4.4
openai==2.32.0
openpyxl==3.1.5
orjson==3.11.8
ormsgpack==1.12.2
packaging==24.2
pandas==3.0.2
pdfminer.six==20251230
pdfplumber==0.11.9
phonenumbers==9.0.28
pillow==12.2.0
pinecone==7.3.0
pinecone-client==6.0.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
preshed==3.0.13
presidio_analyzer==2.2.359
presidio_anonymizer==2.2.362
propcache==0.4.1
protobuf==7.34.1
pyarrow==24.0.0
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
pydeck==0.9.2
Pygments==2.20.0
pypdfium2==5.7.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-file==3.0.1
requests-toolbelt==1.0.0
rich==15.0.0
rpds-py==0.30.0
setuptools==82.0.1
shellingham==1.5.4
simsimd==6.5.16
six==1.17.0
smart_open==7.6.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
spacy==3.8.13
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.3
starlette==1.0.0
streamlit==1.56.0
tenacity==9.1.4
thinc==8.3.13
tiktoken==0.12.0
tldextract==5.3.1
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.45.0
wasabi==1.1.3
watchdog==6.0.0
weasel==1.0.0
wrapt==2.1.2
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
